Add stacktrace logging (#17)
* Add stacktrace logging

The java services are suppressing their stack traces making it
extremely difficult to troubleshoot issues that occur when running
the services locally or in production. The log configurations
lives with the logback.xml

We use a encoder from logstash (https://github.com/logstash/logstash-logback-encoder\#encoders--layouts)
to generated json formatted logs. There is a built in stacktrace
formatter net.logstash.logback.stacktrace.ShortenedThrowableConverter we are
utilising to format the message.

* Remove max length